requirements.txt
Django>=4.0,<4.1
django-import-export
protobuf<4.22
psycopg2>=2.9
